Restructure fixtures for discoverability
