+++ test/conftest.py
@@ -2,12 +2,13 @@
 Fixtures for testing titanium-white package
 """
 
+import json
 import os
+from pathlib import Path
+from typing import Union
 
+from loguru import logger
 import pytest
-
-
-pytest_plugins = ["test.fixtures"]
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -17,3 +18,31 @@ def mock_courtlistener_api_key() -> None:
     """
     courtlistener_env = "COURTLISTENER"
     os.environ[courtlistener_env] = "MOCK_COURTLISTENER_API_KEY"
+
+
+def load_response(response_path: Union[str, Path]) -> dict:
+    """
+    Generic method for handling the loading of the JSON
+    response data from the endpoint
+    """
+    try:
+        response_obj = json.loads(str(response_path))
+    except json.decoder.JSONDecodeError as json_decode_err:
+        json_err_msg = f"Unable to load {str(response_path)} internally as JSON object"
+        logger.error(json_err_msg)
+        raise json_decode_err
+    return response_obj
+
+
+@pytest.fixture(scope="session")
+def people_endpoint_response() -> dict:
+    """
+    Fixture for loading example data from specific endpoints
+    from the courtlistener API
+
+    Loads the response produced by hitting the following endpoitn
+    https://www.courtlistener.com/api/rest/v3/people/?name_first=Cynthia&name_last=Freeland
+    """
+    endpoint_path = Path("./data/test_people_endpoint_response.json").absolute()
+    endpoint_response = load_response(endpoint_path)
+    return endpoint_response
