+++ test/data/test_people_endpoint_response.json
@@ -151,5 +151,3 @@
         }
     ]
 }
-
-
