+++ test/test_courtlistener.py
@@ -13,9 +13,8 @@ def test_empty_courtlistener_construction():
     """
     Test generating and using a CourtListener instance with no arguments
     """
-    with pytest.raises(KeyError):
-        listener_instance = CourtListener()
-        logger.info(listener_instance)
+    listener_instance = CourtListener()
+    logger.info(listener_instance)
 
 
 def test_courtsession_construction(mock_courtlistener_api_key):
@@ -24,12 +23,13 @@ def test_courtsession_construction(mock_courtlistener_api_key):
     """
     mock_endpoint = "https://courtlistener.fake/api/rest/v3/mock"
     session_instance = CourtSession(
-        endpoint_base=mock_endpoint,
+        endpoint=mock_endpoint,
         api_key=mock_courtlistener_api_key,
         headers={},
         parameters={},
     )
-    assert session_instance.endpoint_base == mock_endpoint
+    logger.info(session_instance)
+    assert session_instance.endpoint == mock_endpoint
     assert session_instance.api_key == mock_courtlistener_api_key
     assert not session_instance.headers
     assert not session_instance.parameters
@@ -43,12 +43,13 @@ def test_courtsession_and_courtlistener(mock_courtlistener_api_key):
     """
     mock_endpoint = "https://courtlistener.fake/api/rest/v3/mock"
     session_instance = CourtSession(
-        endpoint_base=mock_endpoint,
+        endpoint=mock_endpoint,
         api_key=mock_courtlistener_api_key,
         headers={},
         parameters={},
     )
-    assert session_instance.endpoint_base == mock_endpoint
+    logger.info(session_instance)
+    assert session_instance.endpoint == mock_endpoint
     assert session_instance.api_key == mock_courtlistener_api_key
     assert not session_instance.headers
     assert not session_instance.parameters
